review: make uuid an optional feature
