+++ arrow_convert/src/field.rs
@@ -5,7 +5,6 @@ use std::sync::Arc;
 use arrow_buffer::{ArrowNativeType, Buffer, ScalarBuffer};
 use arrow_schema::{DataType, Field};
 use chrono::{NaiveDate, NaiveDateTime};
-use crate::uuid_crate::Uuid;
 
 /// The default field name used when a specific name is not provided.
 pub const DEFAULT_FIELD_NAME: &str = "_item";
@@ -210,15 +209,6 @@ impl ArrowField for Vec<u8> {
     #[inline]
     fn data_type() -> DataType {
         DataType::Binary
-    }
-}
-
-impl ArrowField for uuid::Uuid {
-    type Type = Self;
-
-    #[inline]
-    fn data_type() -> DataType {
-        DataType::FixedSizeBinary(16)
     }
 }
