+++ arrow_convert/src/lib.rs
@@ -19,5 +19,3 @@ pub use arrow_convert_derive::{ArrowDeserialize, ArrowField, ArrowSerialize};
 struct ReadmeDoctests;
 
 mod features;
-
-use uuid as uuid_crate;
